# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Resource Management and Booking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register user after Firebase signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "200":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
...
  /hod/users/{id}/activate:
    patch:
      operationId: hodActivateUser
      tags: [hod]
      summary: Activate user and assign role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateUserBody"
      responses:
        "200":
          description: Activated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /hod/approvals/{id}/decide:
    post:
      operationId: hodDecideApproval
      tags: [hod]
      summary: Approve or reject step-2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionBody"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBookingResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingConflictResponse"
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: My bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: from_date
          in: query
          schema:
            type: string
        - name: to_date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: My bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingDetail"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking detail with approval trail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
  /bookings/{id}/cancel:
    patch:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel my booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ANALYTICS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterUserBody:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        firebase_uid:
          type: string
      required: [first_name, last_name, email, firebase_uid]
    AuthUserResponse:
      type: object
      properties:
        user_id:
          type: integer
        role:
          type: string
        is_active:
          type: boolean
        email:
          type: string
      required: [user_id, role, is_active, email]
...
    User:
      type: object
      properties:
        user_id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        is_active:
          type: boolean
        priority_level:
          type: integer
        department_id:
          type: ["integer", "null"]
        department_name:
          type: ["string", "null"]
        created_at:
          type: string
      required: [user_id, first_name, last_name, email, role, is_active, priority_level]
...
    ActivateUserBody:
      type: object
      properties:
        role:
          type: string
      required: [role]
...
    BookingDetail:
      type: object
      properties:
        booking_id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        resource_id:
          type: integer
        resource_name:
          type: string
        category_name:
          type: string
        status_id:
          type: integer
        status_name:
          type: string
        priority_level:
          type: integer
        date:
          type: string
        start_time:
          type: string
        end_time:
          type: string
        purpose:
          type: ["string", "null"]
        created_at:
          type: string
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRecord"
      required: [booking_id, user_id, user_name, resource_id, resource_name, category_name, status_id, status_name, priority_level, date, start_time, end_time, created_at, approvals]
    ApprovalRecord:
      type: object
      properties:
        approval_id:
          type: integer
        step_number:
          type: integer
        decision:
          type: ["string", "null"]
        approver_name:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        approval_time:
          type: ["string", "null"]
      required: [approval_id, step_number]
...
    TimeSlot:
      type: object
      properties:
        start_time:
          type: string
        end_time:
          type: string
        booking_id:
          type: ["integer", "null"]
        purpose:
          type: ["string", "null"]
      required: [start_time, end_time]
    CreateBookingBody:
      type: object
      properties:
        resource_id:
          type: integer
        date:
          type: string
        start_time:
          type: string
        end_time:
          type: string
        purpose:
          type: ["string", "null"]
      required: [resource_id, date, start_time, end_time]
    CreateBookingResponse:
      type: object
      properties:
        booking_id:
          type: integer
        status:
          type: string
        approval_steps:
          type: integer
      required: [booking_id, status, approval_steps]
    BookingConflictResponse:
      type: object
      properties:
        message:
          type: string
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required: [message, alternatives]
...
    ApprovalDecisionBody:
      type: object
      properties:
        decision:
          type: string
        remarks:
          type: ["string", "null"]
      required: [decision]
...